requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
httptools==0.6.1
idna==3.6
motor==3.3.2
odmantic==1.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
